+++ src/dataset.rs
@@ -1,4 +1,8 @@
-use std::{ffi::CString, ffi::NulError, path::Path, ptr};
+use std::{
+    ffi::{CStr, CString, NulError},
+    path::Path,
+    ptr,
+};
 
 use gdal_sys::{CPLErr, GDALDatasetH, GDALMajorObjectH};
 
@@ -10,6 +14,29 @@ use crate::utils::{_last_cpl_err, _last_null_pointer_err, _path_to_c_string, _st
 use crate::{
     gdal_major_object::MajorObject, spatial_ref::SpatialRef, Driver, GeoTransform, Metadata,
 };
+
+pub struct DatasetCapability(&'static CStr);
+
+/// Dataset capabilities
+impl DatasetCapability {
+    /// Dataset can create new layers.
+    pub const CREATE_LAYER: DatasetCapability = DatasetCapability(c"CreateLayer");
+    /// Dataset can delete existing layers.
+    pub const DELETE_LAYER: DatasetCapability = DatasetCapability(c"DeleteLayer");
+    /// Layers of this datasource support CreateGeomField() just after layer creation.
+    pub const CREATE_GEOM_FIELD_AFTER_CREATE_LAYER: DatasetCapability =
+        DatasetCapability(c"CreateGeomFieldAfterCreateLayer");
+    /// Dataset supports curve geometries.
+    pub const CURVE_GEOMETRIES: DatasetCapability = DatasetCapability(c"CurveGeometries");
+    /// Dataset supports (efficient) transactions.
+    pub const TRANSACTIONS: DatasetCapability = DatasetCapability(c"Transactions");
+    /// Dataset supports transactions through emulation.
+    pub const EMULATED_TRANSACTIONS: DatasetCapability = DatasetCapability(c"EmulatedTransactions");
+    /// Dataset has a dedicated GetNextFeature() implementation, potentially returning features from layers in a non sequential way.
+    pub const RANDOM_LAYER_READ: DatasetCapability = DatasetCapability(c"RandomLayerRead");
+    /// Dataset supports calling CreateFeature() on layers in a non sequential way.
+    pub const RANDOM_LAYER_WRITE: DatasetCapability = DatasetCapability(c"RandomLayerWrite");
+}
 
 /// Wrapper around a [`GDALDataset`][GDALDataset] object.
 ///
@@ -327,6 +354,10 @@ impl Dataset {
         }
         Ok(transformation)
     }
+
+    pub fn has_capability(&self, capability: DatasetCapability) -> bool {
+        unsafe { gdal_sys::GDALDatasetTestCapability(self.c_dataset(), capability.0.as_ptr()) == 1 }
+    }
 }
 
 impl MajorObject for Dataset {
@@ -351,7 +382,8 @@ impl Drop for Dataset {
 mod tests {
     use gdal_sys::GDALAccess;
 
-    use crate::test_utils::fixture;
+    use crate::dataset::DatasetCapability;
+    use crate::test_utils::{fixture, open_gpkg_for_update};
     use crate::GdalOpenFlags;
 
     use super::*;
@@ -444,6 +476,19 @@ mod tests {
             },
         )
         .unwrap_err();
+    }
+
+    #[test]
+    fn test_dataset_capabilities() {
+        let ds = Dataset::open(fixture("poly.gpkg")).unwrap();
+        assert!(!ds.has_capability(DatasetCapability::CREATE_LAYER));
+        assert!(!ds.has_capability(DatasetCapability::DELETE_LAYER));
+        assert!(ds.has_capability(DatasetCapability::TRANSACTIONS));
+
+        let (_tmp_path, ds) = open_gpkg_for_update(&fixture("poly.gpkg"));
+        assert!(ds.has_capability(DatasetCapability::CREATE_LAYER));
+        assert!(ds.has_capability(DatasetCapability::DELETE_LAYER));
+        assert!(ds.has_capability(DatasetCapability::TRANSACTIONS));
     }
 
     #[test]